feat: pss-del-set-001 - Retarget remaining imports off transitional Operator Settings public paths
