+++ pkg/services/factory_sessions/internal/runtimeopening/testmain_test.go
@@ -3,9 +3,10 @@ package runtimeopening
 import (
 	"testing"
 
-	_ "github.com/portpowered/infinite-you/pkg/services/operator_settings/testlink"
+	settingswire "github.com/portpowered/infinite-you/pkg/services/operator_settings/wire"
 )
 
 func TestMain(m *testing.M) {
+	settingswire.RegisterTestComposition()
 	m.Run()
 }
